Move win32 alias tokens out
