+++ packages/theming/theme-tokens/src/index.ts
@@ -1,4 +1,3 @@
 export { default as globalTokens } from './tokens-global';
 export { getAliasTokens } from './getAliasTokens';
-export { getOfficeAliasTokens } from './getOfficeAliasTokens';
 export { getMacOSAliasTokens } from './getMacOSAliasTokens';
