+++ packages/theming/win32-theme/src/colorful/tokens-alias.ts
