+++ packages/theming/win32-theme/src/darkGray/tokens-alias.ts
